(BSR)[API] refactor: move collective methods into educational api
