+++ api/src/pcapi/core/educational/exceptions.py
@@ -131,3 +131,7 @@ class InvalidInterventionArea(Exception):
     def __init__(self, errors: list[str]):
         self.errors = errors
         super().__init__()
+
+
+class NoCollectiveBookingToCancel(Exception):
+    pass
